+++ src/android/IonicDeploy.java
@@ -160,7 +160,7 @@ public class IonicDeploy extends CordovaPlugin {
     final SharedPreferences prefs = this.prefs;
 
     if (action.equals("initialize")) {
-      // No need to do anything here.
+      this.server = args.getString(1);
       return true;
     } else if (action.equals("check")) {
       logMessage("CHECK", "Checking for updates");
@@ -195,9 +195,75 @@ public class IonicDeploy extends CordovaPlugin {
     } else if (action.equals("info")) {
       this.info(callbackContext);
       return true;
+    } else if (action.equals("getVersions")) {
+      callbackContext.success(this.getDeployVersions());
+      return true;
+    } else if (action.equals("deleteVersion")) {
+      final String uuid = args.getString(1);
+      boolean status = this.removeVersion(uuid);
+      if (status) {
+        callbackContext.success();
+      } else {
+        callbackContext.error("Error attempting to remove the version, are you sure it exists?");
+      }
+      return true;
+    } else if (action.equals("getMetadata")) {
+      String uuid = null;
+      try {
+        uuid = args.getString(1);
+      } catch (JSONException e) {
+        uuid = this.prefs.getString("upstream_uuid", "");
+      }
+
+      if (uuid.equals("null")) {
+        uuid = this.prefs.getString("upstream_uuid", "");
+      }
+
+      if(uuid == null || uuid.equals("")) {
+        callbackContext.error("NO_DEPLOY_UUID_AVAILABLE");
+      } else {
+        final String metadata_uuid = uuid;
+        this.getMetadata(callbackContext, metadata_uuid);  
+      }
+      return true;
     } else {
       return false;
     }
+  }
+
+  private JSONObject getMetadata(CallbackContext callbackContext, final String uuid) {
+    String endpoint = "/api/v1/apps/" + this.app_id + "/updates/" + uuid + "/";
+    JsonHttpResponse response = new JsonHttpResponse();
+    JSONObject json = new JSONObject();
+    HttpURLConnection urlConnection = null;
+
+    String result = "{}";
+    try {
+      URL url = new URL(this.server + endpoint);
+      urlConnection = (HttpURLConnection) url.openConnection();
+      InputStream in = new BufferedInputStream(urlConnection.getInputStream());
+      result = readStream(in);
+    } catch (MalformedURLException e) {
+      callbackContext.error("DEPLOY_HTTP_ERROR");
+      response.error = true;
+    } catch (IOException e) {
+      callbackContext.error("DEPLOY_HTTP_ERROR");
+      response.error = true;
+    }
+
+    if (urlConnection != null) {
+      urlConnection.disconnect();
+    }
+
+    JSONObject jsonResponse = null;
+    try {
+      jsonResponse = new JSONObject(result);
+      callbackContext.success(jsonResponse);
+    } catch (JSONException e) {
+      response.error = true;
+      callbackContext.error("There was an error fetching the metadata");
+    }
+    return jsonResponse;
   }
 
   private void info(CallbackContext callbackContext) {
@@ -297,6 +363,17 @@ public class IonicDeploy extends CordovaPlugin {
     return prefs.getStringSet("my_versions", new HashSet<String>());
   }
 
+
+  private JSONArray getDeployVersions() {
+    Set<String> versions = this.getMyVersions();
+    JSONArray deployVersions = new JSONArray();
+    for (String version : versions) {
+      String[] version_string = version.split("\\|");
+      deployVersions.put(version_string[0]);
+    }
+    return deployVersions;
+  }
+
   /**
    * Check to see if we already have the version to be downloaded
    *
@@ -370,23 +447,47 @@ public class IonicDeploy extends CordovaPlugin {
     }
   }
 
+  private void removeVersionFromPreferences(String uuid) {
+    SharedPreferences prefs = this.prefs;
+    Set<String> versions = this.getMyVersions();
+    Set<String> newVersions = new HashSet<String>();
+
+    for (String version : versions) {
+      String[] version_string = version.split("\\|");
+      String tempUUID = version_string[0];
+      if (!tempUUID.equals(uuid)) {
+        newVersions.add(version);
+      }
+      prefs.edit().putStringSet("my_versions", newVersions).apply();
+    }
+  }
+
+
   /**
-   * Ugly, lazy bit of code to whack old version directories...
+   * Remove a deploy version from the device
    *
    * @param uuid
+   * @return boolean Success or failure
    */
-  private void removeVersion(String uuid) {
+  private boolean removeVersion(String uuid) {
+    if (uuid.equals(this.getUUID())) {
+      SharedPreferences prefs = this.prefs;
+      prefs.edit().putString("uuid", "").apply();
+      prefs.edit().putString("loaded_uuid", "").apply();
+    }
     File versionDir = this.myContext.getDir(uuid, Context.MODE_PRIVATE);
-
     if (versionDir.exists()) {
       String deleteCmd = "rm -r " + versionDir.getAbsolutePath();
       Runtime runtime = Runtime.getRuntime();
       try {
         runtime.exec(deleteCmd);
+        removeVersionFromPreferences(uuid);
+        return true;
       } catch (IOException e) {
         logMessage("REMOVE", "Failed to remove " + uuid + ". Error: " + e.getMessage());
       }
     }
+    return false;
   }
 
   private JsonHttpResponse postDeviceDetails(String uuid, final String channel_tag) {