+++ src/ios/IonicDeploy.h
@@ -18,6 +18,8 @@
 
 - (void) initVersionChecks;
 
+- (void) initialize:(CDVInvokedUrlCommand *)command;
+
 - (void) check:(CDVInvokedUrlCommand *)command;
 
 - (void) download:(CDVInvokedUrlCommand *)command;
@@ -32,13 +34,23 @@
 
 - (NSMutableArray *) getMyVersions;
 
+- (NSMutableArray *) getDeployVersions;
+
+- (void) getVersions:(CDVInvokedUrlCommand *)command;
+
+- (void) getMetadata:(CDVInvokedUrlCommand *)command;
+
 - (bool) hasVersion:(NSString *) uuid;
 
 - (void) saveVersion:(NSString *) uuid;
 
+- (void) deleteVersion:(CDVInvokedUrlCommand *)command;
+
 - (void) cleanupVersions;
 
-- (void) removeVersion:(NSString *) uuid;
+- (BOOL) removeVersion:(NSString *) uuid;
+
+- (BOOL) excludeVersionFromBackup:(NSString *) uuid;
 
 @end