+++ changelog.md
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+## 0.4.0
+
+* Added `getMetadata` method to fetch deploy metadata
+* Added `getVersions` and `removeVersion` methods. They will allow you to manage the deploys 
+  currently on the device.
+* (fix) iOS deploys will now give a download error if the app goes into the background while
+  downloading a deploy.
+* (fix) Excluding deploys from iOS cloud backups
+* (fix) iOS rollbacks now behave the same as android
+
+
 ## 0.3.0
 
 * Changed plugin id from `com.ionic.deploy` to `ionic-plugin-deploy`